+++ R/get_collection.R
@@ -1,17 +1,15 @@
 #' Get a collection of networks
 #'
 #' Retrieve a set of networks based on the results of one of the `search_*()`
-#' function. The function also accept a numeric vector of Mangal network IDs.
+#' function. The function also accepts a numeric vector of Mangal network IDs.
 #'
 #' @param x `numeric` vector of Mangal network IDs or an object returned by
 #' by one of the `search_*()` functions.
 #' @param ... arguments to be passed on to [rmangal::get_network_by_id()].
 #'
 #' @return
-#'  If there is only one network to be retrieved, `get_collection()` returns a
-#' `mgNetwork` object, otherwise it returns a object of class
-#' `mgNetworksCollection` which is a collection (a list) of `mgNetwork`
-#' objects [rmangal::get_network_by_id()]).
+#'  Returns a object of class `mgNetworksCollection` which is a collection 
+#' (actually, a list) of `mgNetwork` objects [rmangal::get_network_by_id()]).
 #'
 #' @seealso
 #' [search_datasets()], [search_interactions()], [search_networks()],
@@ -19,8 +17,8 @@
 #'
 #' @examples
 #' \donttest{
-#' mg_2 <- get_collection(c(1076:1077), verbose = FALSE)
-#' mg_anemone <- get_collection(search_networks(query='anemone%'), verbose = FALSE)
+#'  mg_2 <- get_collection(c(1076:1077), verbose = FALSE)
+#'  mg_anemone <- get_collection(search_networks(query='anemone%'), verbose = FALSE)
 #' }
 #' @export
 
@@ -31,7 +29,7 @@ get_collection <- function(x, ...) {
 #' @describeIn get_collection Get a collection of networks (default).
 #' @export
 get_collection.default <- function(x, ...) {
-  get_network_by_id(x, ...)
+  get_network_by_id(x, force_collection = TRUE, ...)
 }
 
 #' @describeIn get_collection Get a collection of networks from a `mgSearchDatasets` object.
@@ -39,14 +37,14 @@ get_collection.default <- function(x, ...) {
 get_collection.mgSearchDatasets <- function(x, ...) {
   # Get networks ids
   net_ids <- unique(unlist(purrr::map(x$networks, "id")))
-  get_network_by_id(net_ids, ...)
+  get_collection.default(net_ids, ...)
 }
 
 #' @describeIn get_collection Get a collection of networks from a `mgSearchNetworks` object.
 #' @export
 get_collection.mgSearchNetworks <- function(x, ...) {
   # Get networks ids
-  get_network_by_id(x$id, ...)
+  get_collection.default(x$id, ...)
 }
 
 
@@ -55,7 +53,7 @@ get_collection.mgSearchNetworks <- function(x, ...) {
 get_collection.mgSearchReferences <- function(x, ...) {
   # Get networks ids
   net_ids <- unique(unlist(purrr::map(x$networks, "id")))
-  get_network_by_id(net_ids, ...)
+  get_collection.default(net_ids, ...)
 }
 
 #' @describeIn get_collection Get a collection of networks from a `mgSearchNodes` object.
@@ -63,7 +61,7 @@ get_collection.mgSearchReferences <- function(x, ...) {
 get_collection.mgSearchNodes <- function(x, ...) {
   # Get networks ids
   net_ids <- unique(x$network_id)
-  get_network_by_id(net_ids, ...)
+  get_collection.default(net_ids, ...)
 }
 
 #' @describeIn get_collection Get a collection of networks from a `mgSearchTaxa` object.
@@ -71,7 +69,7 @@ get_collection.mgSearchNodes <- function(x, ...) {
 get_collection.mgSearchTaxonomy <- function(x, ...) {
   # Get networks ids
   net_ids <- unique(x$network_id)
-  get_network_by_id(net_ids, ...)
+  get_collection.default(net_ids, ...)
 }
 
 
@@ -80,5 +78,5 @@ get_collection.mgSearchTaxonomy <- function(x, ...) {
 get_collection.mgSearchInteractions <- function(x, ...) {
   # Get networks ids
   net_ids <- unique(x$network_id)
-  get_network_by_id(net_ids, ...)
+  get_collection.default(net_ids, ...)
 }