+++ R/search_networks.R
@@ -6,7 +6,7 @@
 #' partial match was found are returned.
 #' Alternatively, a named list can be used to look for an exact match in a specific column (see Details section)
 #'
-#' @param query either a character string including a single keyword or a list containing a custom query (see details section below), or a spatial object (see the description of `query_sf`).
+#' @param query either a character string including a single keyword or a named list containing a custom query (see details section below), or a spatial object (see the description of `query_sf`).
 #' Note that if an empty character string is passed, then all datasets available are returned.
 #' @param query_sf a spatial object of class `sf` used to search in a specific geographical area.
 #' @param verbose a `logical`. Should extra information be reported on progress?
@@ -24,26 +24,26 @@
 #' - public: network publicly available
 #'
 #' Note that for lists with more than one element, only the first element is used, the others are ignored. An example is provided below.
-#'
 #' @references
-#' Metadata available at <https://mangal-wg.github.io/mangal-api/#networks>
+#' * <https://mangal.io/#/>
+#' * <https://mangal-interactions.github.io/mangal-api/#networks>
 #'
 #' @examples
-#' mg_insect <- search_networks(query="insect%")
 #' \donttest{
-#' # Retrieve the search results
-#' nets_insect <- get_collection(mg_insect)
-#' # Spatial query
-#' library(sf)
-#' library(USAboundaries)
-#' area <- us_states(state="california")
-#' networks_in_area <- search_networks_sf(area, verbose = FALSE)
-#' plot(networks_in_area)
+#'  mg_insect <- search_networks(query="insect%")
+#'  # Retrieve the search results
+#'  nets_insect <- get_collection(mg_insect)
+#'  # Spatial query
+#'  library(sf)
+#'  library(USAboundaries)
+#'  area <- us_states(state="california")
+#'  networks_in_area <- search_networks_sf(area, verbose = FALSE)
+#'  plot(networks_in_area)
+#'  # Retrieve network ID 5013
+#'  net_5013 <- search_networks(query = list(id = 5013))
+#'  # Network(s) of dataset ID 19
+#'  mg_19 <- search_networks(list(dataset_id = 19))
 #' }
-#' # Retrieve network ID 5013
-#' net_5013 <- search_networks(query = list(id = 5013))
-#' # Network(s) of dataset ID 19
-#' mg_19 <- search_networks(list(dataset_id = 19))
 #'
 #' @export
 
@@ -70,7 +70,7 @@ search_networks <- function(query, verbose = TRUE, ...) {
 #' @export
 search_networks_sf <- function(query_sf, verbose = TRUE, ...) {
 
-  stopifnot("sf" %in% class(query_sf))
+  stopifnot(is(query_sf, "sf"))
   stop_if_missing_sf()
 
   # API doesn't allow spatial search yet, so we call sf