+++ R/search_nodes.R
@@ -6,7 +6,7 @@
 #' partial match was found are returned.
 #' Alternatively, a named list can be used to look for an exact match in a specific column (see Details section)
 #'
-#' @param query either a character string including a single keyword or a list containing a custom query (see details section below).
+#' @param query either a character string including a single keyword or a named list containing a custom query (see details section below).
 #' Note that if an empty character string is passed, then all datasets available are returned.
 #' @param verbose a `logical`. Should extra information be reported on progress?
 #' @param ... further arguments to be passed to [httr::GET()].
@@ -29,12 +29,14 @@
 #' All networks in which taxa are involved are also attached to the `data.frame`.
 #'
 #' @references
-#' Metadata available at <https://mangal-wg.github.io/mangal-api/#nodes>
+#' * <https://mangal.io/#/>
+#' * <https://mangal-interactions.github.io/mangal-api/#nodes>
 #'
 #' @examples
-#' res_acer <- search_nodes("Acer")
-#' res_926 <- search_nodes(list(network_id = 926))
-#'
+#' \donttest{
+#'  res_acer <- search_nodes("Acer")
+#'  res_926 <- search_nodes(list(network_id = 926))
+#' }
 #' @export
 
 search_nodes <- function(query, verbose = TRUE, ...) {
@@ -43,7 +45,8 @@ search_nodes <- function(query, verbose = TRUE, ...) {
 
   nodes <- resp_to_df_flt(get_gen(endpoints()$node, query = query,
       verbose = verbose, ...)$body)
-  if (all(dim(nodes) == 1)) return(data.frame())
+  if (all(dim(nodes) == 1)) 
+    return(data.frame())
 
   nodes <- nodes[, names(nodes)[names(nodes) != "taxonomy"]]