+++ R/search_taxonomy.R
@@ -9,18 +9,18 @@
 #'
 #' @param query a character string including a single keyword.
 #' Note that if an empty character string is passed, then all datasets available are returned.
-#' @param tsn a `numeric`. Unique taxonomic identifier from Integrated Taxonomic Information System (<https://www.itis.gov/>).
-#' @param gbif a `numeric`. Unique taxonomic identifier from Global Biodiversity Information Facility (<https://www.gbif.org/>).
-#' @param eol a `numeric`. Unique taxonomic identifier from Encyclopedia of Life (<https://eol.org/>).
-#' @param col a `numeric`. Unique taxonomic identifier from Catalogue of Life (<https://www.catalogueoflife.org/>).
-#' @param bold a `numeric`. Unique taxonomic identifier from Barcode of Life (<http://www.boldsystems.org/>).
-#' @param ncbi a `numeric`. Unique taxonomic identifier from National Center for Biotechnology Information (<https://www.ncbi.nlm.nih.gov/>).
+#' @param tsn a `numeric`. Unique taxonomic identifier from Integrated Taxonomic Information System (<https://www.itis.gov>).
+#' @param gbif a `numeric`. Unique taxonomic identifier from Global Biodiversity Information Facility (<https://www.gbif.org>).
+#' @param eol a `numeric`. Unique taxonomic identifier from Encyclopedia of Life (<https://eol.org>).
+#' @param col a `numeric`. Unique taxonomic identifier from Catalogue of Life (<https://www.catalogueoflife.org>).
+#' @param bold a `numeric`. Unique taxonomic identifier from Barcode of Life (<http://www.boldsystems.org>).
+#' @param ncbi a `numeric`. Unique taxonomic identifier from National Center for Biotechnology Information (<https://www.ncbi.nlm.nih.gov>).
 #' @param verbose a `logical`. Should extra information be reported on progress?
 #' @param ... further arguments to be passed to [httr::GET()].
 #'
 #' @details
 #' Taxon names of the `taxonomy` table were validated with
-#' TNRS (see <http://tnrs.iplantcollaborative.org/> and/or GNR
+#' TNRS (see <https://tnrs.biendata.org> and/or GNR
 # ' (see <https://resolver.globalnames.org/>). The taxon names in this table
 #' might not be the taxon name documented in the original publication.
 #' In order to identify relevant networks with the original name, use
@@ -29,22 +29,25 @@
 #' The validation of taxon names was performed by an automated
 #' procedure and if there is any doubt, the original names recorded
 #' by authors should be regarded as the most reliable information. Please
-#' report any issue related to taxonomy at <https://github.com/mangal-wg/mangal-datasets/issues>.
+#' report any issue related to taxonomy at <https://github.com/mangal-interactions/contribute/issues/new/choose/>.
 #'
 #' @return
 #' An object of class `mgSearchTaxonomy`, which is a `data.frame` including
 #' all taxa matching the query.
 #'
 #' @references
-#' Metadata available at <https://mangal-wg.github.io/mangal-api/#taxonomy>
+#' * <https://mangal.io/#/>
+#' * <https://mangal-interactions.github.io/mangal-api/#taxonomy>
 #'
 #' @seealso
 #' [search_nodes()]
 #'
 #' @examples
-#' search_taxonomy("Acer")
-#' # Retrieve higher classification
-#' tsn_acer <- search_taxonomy("Acer")$taxonomy.tsn
+#' \donttest{
+#'  search_taxonomy("Acer")
+#'  # Retrieve higher classification
+#'  tsn_acer <- search_taxonomy("Acer")$taxonomy.tsn
+#' }
 #' @export
 
 search_taxonomy <- function(query, tsn = NULL, gbif = NULL, eol = NULL,
@@ -57,7 +60,7 @@ search_taxonomy <- function(query, tsn = NULL, gbif = NULL, eol = NULL,
 
   if (length(req)) {
       if (length(req) > 1) {
-        stop("Query with multiple criteria not allowed.")
+        stop("Queries with multiple criteria are not allowed.")
       } else query <- req
   } else {
     stopifnot(is.character(query))