+++ tests/testthat/test-get_citation.R
@@ -1,11 +1,9 @@
-context("get_citation")
-
-ref <- get_citation(get_network_by_id(18))
-netws <- get_collection(search_networks("lagoon"))
-ref2 <- get_citation(netws)
-
-
 test_that("expected behavior", {
+  vcr::use_cassette(name = "get_citation", {
+    ref <- get_citation(get_network_by_id(18))
+    netws <- get_collection(search_networks("lagoon"))
+    ref2 <- get_citation(netws)
+  })
   expect_equal(length(ref), 1)
   expect_equal(length(ref2), 2)
   expect_true(grepl("^@article", ref))